+++ src/qanDraggableCtrl.cpp
@@ -242,32 +242,36 @@ void    DraggableCtrl::dragMove(const QPointF& sceneDragPos, bool dragSelection)
     // 2.2. If target position is "centered" on grid
     //    or mouse delta > grid
     //   2.2.1 Compute snapped position, apply it
+    const auto targetDragOrientation = _targetItem->getDragOrientation();
+    const auto dragHorizontally = (targetDragOrientation == qan::NodeItem::DragOrientation::All) ||
+                                (targetDragOrientation == qan::NodeItem::DragOrientation::Horizontal);
+    const auto dragVertically = (targetDragOrientation == qan::NodeItem::DragOrientation::All) ||
+                                (targetDragOrientation == qan::NodeItem::DragOrientation::Vertical);
     if (getGraph()->getSnapToGrid()) {
         const auto& gridSize = getGraph()->getSnapToGridSize();
-        bool applyX = std::fabs(delta.x()) > (gridSize.width() / 2.001);
-        bool applyY = std::fabs(delta.y()) > (gridSize.height() / 2.001);
-
-        //qWarning() << "--------";
-        //qWarning() << "targetUnsnapPos=" << targetUnsnapPos;
-        if (!applyX) {
+        bool applyX = dragHorizontally &&
+                      std::fabs(delta.x()) > (gridSize.width() / 2.001);
+        bool applyY = dragVertically &&
+                      std::fabs(delta.y()) > (gridSize.height() / 2.001);
+        if (!applyX && dragHorizontally) {
             const auto posModGridX = fmod(targetUnsnapPos.x(), gridSize.width());
-            //qWarning() << "posModGridX=" << posModGridX;
             applyX = qFuzzyIsNull(posModGridX);
         }
-        if (!applyY) {
+        if (!applyY && dragVertically) {
             const auto posModGridY = fmod(targetUnsnapPos.y(), gridSize.height());
-            //qWarning() << "posModGridY=" << posModGridY;
             applyY = qFuzzyIsNull(posModGridY);
         }
-        //qWarning() << "applyX=" << applyX << "   applyY=" << applyY;
         if (applyX || applyY) {
-            const auto targetSnapPosX = gridSize.width() * std::round(targetUnsnapPos.x() / gridSize.width());
-            const auto targetSnapPosY = gridSize.height() * std::round(targetUnsnapPos.y() / gridSize.height());
+            const auto targetSnapPosX = dragHorizontally ? gridSize.width() * std::round(targetUnsnapPos.x() / gridSize.width()) :
+                                                           _initialTargetPos.x();
+            const auto targetSnapPosY = dragVertically ? gridSize.height() * std::round(targetUnsnapPos.y() / gridSize.height()) :
+                                                         _initialTargetPos.y();
             _targetItem->setPosition(QPointF{targetSnapPosX,
                                              targetSnapPosY});
         }
     } else { // Do not snap to grid
-        _targetItem->setPosition(targetUnsnapPos);
+        _targetItem->setPosition(QPointF{dragHorizontally ? targetUnsnapPos.x() : _initialTargetPos.x(),
+                                         dragVertically   ? targetUnsnapPos.y() : _initialTargetPos.y()});
     }
     // FIXME #185 Selection move does not works...
     if (dragSelection) {